moved vault man into sub pkg
