+++ src/main2.py
@@ -10,7 +10,7 @@ import subprocess as sp
 from pathlib import Path
 
 from libr3dv1t.central_config import default_rvcc as _rvcc
-from libr3dv1t.vault_man import VaultMan
+from libr3dv1t.vault.vault_man import VaultMan
 from libr3dv1t.typedefs import VaultObj
 from libr3dv1t.errors import R3D_IO_Error
